Moved configured box into a separated file
