+++ hivez/lib/src/builders/box_creator.dart
@@ -75,14 +75,3 @@ class BoxCreatorImpl implements BoxCreator {
     };
   }
 }
-
-class ConfiguredBox<K, T> extends BoxDecorator<K, T> {
-  final BoxConfig config;
-
-  ConfiguredBox(
-    this.config,
-  ) : super(config.createBox<K, T>());
-
-  BoxType get type => config.type;
-  LogHandler? get logger => config.logger;
-}
