+++ hivez/lib/src/builders/builders.dart
@@ -5,3 +5,4 @@ import 'package:hivez/src/boxes/boxes.dart';
 
 part 'box_creator.dart';
 part 'box_config.dart';
+part 'configured_box.dart';
